added different files for pages and optimised the code
