+++ eslint.config.mjs
@@ -13,4 +13,5 @@ const eslintConfig = [
   ...compat.extends("next/core-web-vitals", "next/typescript"),
 ];
 
+
 export default eslintConfig;
